LayoutTests/fast/js/regexp-extended-characters-match.html

<html>
<body>
<h3>This test checks regular expressions using extended (> 255) characters and character classes.</h3>
<hr>
<script>
function test(s)
{
    var result = eval(s);
    document.write("<p>" + s + ": " + result + "</p>");
}

if (window.layoutTestController)
    layoutTestController.dumpAsText();

document.write("<h3>These regular expressions should return null:</h3>");

//test('var r = new RegExp("[\u0101-\u0100]"); r.exec("a")');

test('var r = new RegExp("[\u0100-\u0101]"); r.exec("a")');

test('var r = new RegExp("[\u0100]"); r.exec("a")');

test('var r = new RegExp("\u0100"); r.exec("a")');

document.write("<h3>These regular expressions should return \"a\"</h3>");

test('var r = new RegExp("[\u0061]"); r.exec("a")');

test('var r = new RegExp("[\u0100-\u0101a]"); r.exec("a")');

test('var r = new RegExp("[\u0100a]"); r.exec("a")');

test('var r = new RegExp("\u0061"); r.exec("a")');

test('var r = new RegExp("[a-\u0100]"); r.exec("a")');

document.write("<h3>These regular expressions should return \"\u0100\":</h3>");

test('var r = new RegExp("[\u0100]"); r.exec("\u0100")');

test('var r = new RegExp("[\u0100-\u0101]"); r.exec("\u0100")');

test('var r = new RegExp("\u0100"); r.exec("\u0100")');

</script>
</body>
</html>
